Correção nulls

+++ Grupo2A/Model/DTO/Prato2listing_dto.cs
@@ -2,7 +2,7 @@ namespace Cozinha_BE.Model.DTO;
 
 public class Prato2listing_dto{
 
-    public int IdPrato {get; set; }
+    public long IdPrato {get; set; }
     public required string Nome {get; set; }
     public required TipoDePrato TipoPrato {get; set;}
     public required bool Ativo { get; set; }
